+++ vocabularies/scripts/validate-changes.py
@@ -1,4 +1,5 @@
-# This script compares ../json-ld/locations.json in the current branch to that in `master`
+# This script compares ../json-ld/locations.json (or other named files) in
+# the current branch to that in `master`
 #
 # Depends on Python 3
 #
@@ -7,15 +8,26 @@
 #
 # Usage:
 #   git checkout [feature-branch]
-#   python validate-locations.py
+#   python validate-changes.py [which]
+#
+#   Params:
+#    - [which]: Name of json-ld file to compare
+#               (e.g. 'recapCustomerCodes', 'organizations').
+#               Default 'locations'
 
 from deepdiff import DeepDiff
 from git import Git
 import json
+import sys
 
-# Provides a comparison on the serialized Locations JSON-LD object in the
+# Provides a comparison on the serialized named JSON-LD object in the
 # current working branch
 def main():
+    # Determine what json-ld file to compare
+    which = 'locations'
+    if len(sys.argv) == 2:
+        which = sys.argv[1]
+
     # Initialize git client and store current branch name
     git = Git()
     currentBranch = None
@@ -24,27 +36,27 @@ def main():
             currentBranch = branch[2:]
             break
 
-    print(f'Comparing {currentBranch} to master')
+    print(f'Comparing {which} in {currentBranch} to master')
 
-    # Get current working copy of the locations JSON-LD file
-    newLocationFile = openLocationsJsonFile()
+    # Get current working copy of the JSON-LD file
+    newFile = openJsonFile(which)
 
     try:
         # Checkout master and get the current version
         git.checkout('master')
-        masterLocationFile = openLocationsJsonFile()
+        masterFile = openJsonFile(which)
 
         # Compare the two objects
-        for item in newLocationFile['@graph']:
+        for item in newFile['@graph']:
             if not 'skos:notation' in item:
                 print('Item found without a skos:notation!', item)
 
-        newLocDict = {item['skos:notation']: item for item in newLocationFile['@graph']}
-        masterLocDict = {item['skos:notation']: item for item in masterLocationFile['@graph']}
+        newLocDict = {item['skos:notation']: item for item in newFile['@graph']}
+        masterLocDict = {item['skos:notation']: item for item in masterFile['@graph']}
         newKeys = newLocDict.keys() - masterLocDict.keys()
         deletedKeys = masterLocDict.keys() - newLocDict.keys()
         alteredKeys = list(filter(lambda x: x[1], [
-            (key, DeepDiff(newLocDict[key], masterLocDict[key], ignore_order=True))
+            (key, DeepDiff(masterLocDict[key], newLocDict[key], ignore_order=True))
             for key in set(newLocDict.keys()) & set(masterLocDict.keys())
         ]))
 
@@ -63,8 +75,8 @@ def main():
     git.checkout(currentBranch)
 
 
-def openLocationsJsonFile():
-    with open('./../json-ld/locations.json') as locFile:
+def openJsonFile(which):
+    with open(f'./../json-ld/{which}.json') as locFile:
         return json.load(locFile)
 
 
@@ -77,12 +89,12 @@ def displayAlterations(alteredKeys):
                     print('Attribute: {}'.format(label))
                     print('Old Value: {}'.format(changes['old_value']))
                     print('New Value: {}'.format(changes['new_value']))
-        
+
         if 'iterable_item_removed' in diff.keys():
             for label, removal in diff['iterable_item_removed'].items():
                 print('Attribute(Pos): {}'.format(label))
                 print('Removed Value: {}'.format(removal))
-        
+
         if 'iterable_item_added' in diff.keys():
             for label, addition in diff['iterable_item_added'].items():
                 print('Attribute(Pos): {}'.format(label))